requirements.txt
tqdm>=4.19.5
numpy>=1.13.0
Pillow>=4.1.1
tensorflow>=1.7.1
six
protobuf